Add the laws implementations
The motivation behind this is to allow any person wanting to implement
any structure that should conform to the specification will also have
the readme and more importantly the implementations of the laws.

The advantage of this as well is that any changes to the spec should be
reflected in the laws folder.